fix: CarouselPage 컴포넌트 Main에 합치기
