+++ src/pages/main/Main.tsx
@@ -1,7 +1,7 @@
 import styled from '@emotion/styled';
+import { Suspense, lazy } from 'react';
 import { useParams } from 'react-router-dom';
 
-import CarouselPage from './components/CarouselPage';
 import FaqDropdown from './components/FaqDropdown';
 import FaqTitle from './components/FaqTitle';
 import Introduction from './components/Introduction';
@@ -9,24 +9,38 @@ import { AuthorizationHeader, UnAuthorizationHeader } from './components/MainHea
 import Manual from './components/Manual';
 import OnBoarding from './components/OnBoarding';
 import Ruler from './components/Ruler';
+import { SkeletonComponent } from './components/skeletons/SkeletonComponent';
 import { FAQ_DATA } from './constants/faqData';
-import { useGetRecommendTopic } from './hooks/queries';
+import { useGetGroupContent, useGetRecommendTopic } from './hooks/queries';
 
 import Footer from './../../components/commons/Footer';
 import Spacing from './../../components/commons/Spacing';
 
 const Main = () => {
+  const lazyCarousel = import('./components/GroupCarousel');
+  const LazyCarousel = lazy(() => lazyCarousel);
   const { content } = useParams();
   const { moimId } = useParams();
   const topic = useGetRecommendTopic(content || '');
+  const { data, groupLength, isLoading } = useGetGroupContent(moimId || '');
+  console.log(data);
 
   return (
     <MainPageWrapper>
       {localStorage.getItem('accessToken') ? <AuthorizationHeader /> : <UnAuthorizationHeader />}
       <OnBoarding />
-      <CarouselLayout>
-        <CarouselPage moimId={moimId} />
-      </CarouselLayout>
+      <CarouselComponentWrapper>
+        <TitleLayout>마일과 함께하고 있는 글 모임이에요</TitleLayout>
+        {isLoading && groupLength ? (
+          <SkeletonComponent groupLength={groupLength} />
+        ) : (
+          groupLength && (
+            <Suspense fallback={<SkeletonComponent groupLength={groupLength} />}>
+              <LazyCarousel data={data} groupLength={groupLength} />
+            </Suspense>
+          )
+        )}
+      </CarouselComponentWrapper>
       <Ruler data={topic?.data} />
       <Introduction />
       <Manual />
@@ -47,15 +61,27 @@ const Main = () => {
 export default Main;
 
 const MainPageWrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
   width: 100%;
 
   background-color: ${({ theme }) => theme.colors.backGroundGray};
 `;
 
-const CarouselLayout = styled.section`
+const CarouselComponentWrapper = styled.section`
   display: flex;
-  align-items: center;
-  justify-content: center;
+  flex-direction: column;
+  width: fit-content;
+  padding-bottom: 10rem;
+`;
+
+const TitleLayout = styled.div`
+  display: flex;
+  padding-top: 7.2rem;
+
+  ${({ theme }) => theme.fonts.title3};
 `;
 
 const FaqLayout = styled.section`
